combine functional into unit tests
